exception page

+++ src/main/java/com/sjsu/cmpe275/controller/ProfileController.java
@@ -8,10 +8,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 /**
  * Created by chenglongwei on 3/24/16.
@@ -21,17 +18,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class ProfileController {
     @Autowired
     ProfileService service;
-    private static final int DEFAULT_PAGE_NUM = 0;
-    private static final int DEFAULT_PAGE_SIZE = 15;
-
-    @RequestMapping(value = "/profile/list")
-    public String list(@RequestParam(value = "page", required = false) Integer page,
-                       Model model) {
-        int pageNum = page != null ? page : DEFAULT_PAGE_NUM;
-        Page<Profile> paging = service.findAll(pageNum, DEFAULT_PAGE_SIZE);
-        model.addAttribute("page", paging);
-        return "profile/list";
-    }
 
     // Update or Delete one user profile.
     @RequestMapping(value = "/profile/{id}", method = RequestMethod.GET)
@@ -57,34 +43,16 @@ public class ProfileController {
 
     // Create user profile.
     @RequestMapping(value = "/profile", method = RequestMethod.GET)
-    public String create() {
+    public String create(Model model) {
+        model.addAttribute("profile", new Profile());
         return "profile/create";
     }
 
     // Insert or update user profile.
     @RequestMapping(value = "/profile/{id}", method = RequestMethod.POST)
-    public String create(@PathVariable("id") long id,
-                         @RequestParam(value = "firstname") String firstname,
-                         @RequestParam(value = "lastname") String lastname,
-                         @RequestParam(value = "address") String address,
-                         @RequestParam(value = "organization") String organization,
-                         @RequestParam(value = "aboutmyself") String aboutmyself,
-                         BindingResult bindingResult,
-                         Model model) {
-        // Check params.
-        if (bindingResult.hasErrors()) {
-            model.addAllAttributes(bindingResult.getModel());
-            String message = "Create or update user with ID " + id + " Error.";
-            throw new ProfileException(message);
-        }
-
-        Profile profile = new Profile();
-        profile.setId(id);
-        profile.setFirstname(firstname);
-        profile.setLastname(lastname);
-        profile.setAddress(address);
-        profile.setOrganization(organization);
-        profile.setAboutMyself(aboutmyself);
+    public String create(
+            @PathVariable("id") long id,
+            @ModelAttribute("profile") Profile profile) {
 
         // Insert or update profile.
         if (service.findById(profile.getId()) == null) {
@@ -93,7 +61,7 @@ public class ProfileController {
             service.update(profile);
         }
 
-        return "redirect:/profile/" + id;
+        return "redirect:/profile/" + profile.getId();
     }
 
     // Delete a profile.
@@ -102,6 +70,11 @@ public class ProfileController {
     public String delete(
             @RequestParam(value = "page", required = false) Integer page,
             @PathVariable("userID") long id) {
+
+        if (service.findById(id) == null) {
+            String message = "Sorry, the requested user with ID " + id + " does not exist.";
+            throw new ProfileException(message);
+        }
 
         service.deleteById(id);
